Refactor some stuffs
